move function into own file
