+++ packages/aila/src/core/chat/AilaStreamHandler.ts
@@ -6,6 +6,7 @@ import { AilaThreatDetectionError } from "../../features/threatDetection/types";
 import { AilaChatError } from "../AilaError";
 import type { AilaChat } from "./AilaChat";
 import type { PatchEnqueuer } from "./PatchEnqueuer";
+import { createStreamReaderFromURL } from "./createStreamReaderFromURL";
 
 const log = aiLogger("aila:stream");
 export class AilaStreamHandler {
@@ -95,8 +96,8 @@ export class AilaStreamHandler {
 
     const lastUserMessage = this._chat.messages.slice(-1)[0];
     if (AILA_ENGINE_STREAM_URL) {
-      this._streamReader = await createStreamReader(
-        `${AILA_ENGINE_STREAM_URL}?prompt=${JSON.stringify(currentLessonPlan)}\n\n\n\n${lastUserMessage?.content}`,
+      this._streamReader = await createStreamReaderFromURL(
+        `${AILA_ENGINE_STREAM_URL}?prompt=${JSON.stringify(currentLessonPlan)}\n\n${lastUserMessage?.content}`,
       );
     } else {
       const messages = this._chat.completionMessages();
@@ -159,53 +160,4 @@ export class AilaStreamHandler {
       this._controller.close();
     }
   }
-}
-
-async function createStreamReader(
-  apiUrl: string,
-): Promise<ReadableStreamDefaultReader<string>> {
-  const response = await fetch(apiUrl, {
-    method: "GET",
-    headers: {
-      Accept: "application/json",
-    },
-  });
-
-  if (!response.ok || !response.body) {
-    throw new Error(`Failed to fetch stream from API: ${response.statusText}`);
-  }
-
-  const textDecoder = new TextDecoder(); // Decodes binary chunks to strings
-  const streamReader = response.body.getReader();
-
-  return new ReadableStream<string>({
-    async start(controller) {
-      const readChunk = async () => {
-        try {
-          const { done, value } = await streamReader.read();
-          if (done) {
-            controller.close();
-            return;
-          }
-
-          // Decode binary chunk and enqueue each line
-          const decoded = textDecoder.decode(value, { stream: true });
-          const lines = decoded.split("\n");
-          for (const line of lines) {
-            if (line.trim()) {
-              controller.enqueue(line.trim());
-            }
-          }
-          await readChunk();
-        } catch (error) {
-          controller.error(error);
-        }
-      };
-
-      await readChunk();
-    },
-    async cancel() {
-      await streamReader.cancel();
-    },
-  }).getReader();
 }
